feature: db.py with add, remove, and load to table. cfknn now ONLY trains/runs inference
